pyproject.toml
[project]
name = "yolo-training-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyqt6>=6.8.1",
    "pyqt6-qt6>=6.8.2",
    "pyqt6-sip>=13.10.0",
    "PyQt6-WebEngine>=6.6.0",
    "ultralytics>=8.3.78",
    "albumentations>=1.3.1",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "flask>=3.0.2",
    "pillow>=10.2.0"
]
